requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
boto3==1.9.61
boto==2.49.0
botocore==1.12.62         # via boto3, s3transfer
django-appconf==1.0.2     # via django-select2
django-countries==5.3.2
django-debug-toolbar==1.10.1
django-parler==1.9.2
django-s3-folder-storage==0.5
django-select2==6.3.1
django-storages==1.7.1
django==2.1.4
docutils==0.14            # via botocore
html5lib==1.0.1
jmespath==0.9.3           # via boto3, botocore
pillow==5.4.1
psycopg2-binary==2.7.6.1
pypdf2==1.26.0
python-dateutil==2.7.5    # via botocore
pytz==2018.5
reportlab==3.5.13
s3transfer==0.1.13        # via boto3
six==1.11.0               # via html5lib, python-dateutil, xhtml2pdf
sqlparse==0.2.4           # via django-debug-toolbar
urllib3==1.24.1           # via botocore
webencodings==0.5.1
xhtml2pdf==0.2.3
